requirements.txt
python-telegram-bot>=21.5
APScheduler>=3.10.4
python-dotenv>=1.0.0
